requirements/local.txt
Django==2.0.7
django-crispy-forms==1.7.2
django-discover-runner==1.0
django-environ==0.4.5
django-role-permissions==2.2.0
Pillow==5.2.0
psycopg2==2.7.5
pytz==2018.5
six==1.11.0
whitenoise==3.3.1
